requirements.txt
Flask
python-dotenv
psycopg2
PyPDF2
pytesseract 